Add NoticeLayout.tsx : 컴포넌트 / 페이지 분리 위함
